Split Oak Functions binaries
The main change is to turn oak_functions/loader into a library instead
of a combined library + binary. This means that most of the code in
main.rs is now in lib.rs, but the actual main is in the separate bin
crates now.

Also

- Fix broken documentation
- Simplify runner

Fixes #2549
