+++ .github/workflows/provenance.yaml
@@ -53,7 +53,7 @@ jobs:
         run: |
           ./scripts/build_provenance \
           "oak_functions" \
-          ./oak_functions/loader/bin/oak_functions_loader \
+          ./target/x86_64-unknown-linux-musl/release/oak_functions_loader_base \
           ./scripts/runner build-functions-server
 
       - name: Move new provenance files to the provenance branch
