+++ oak_functions/examples/mobilenet/example.toml
@@ -11,7 +11,7 @@ target = { Cargo = { cargo_manifest = "oak_functions/examples/mobilenet/module/C
 additional_args = [
   "--config-path=./oak_functions/examples/mobilenet/config.toml"
 ]
-required_features = ["oak-unsafe"]
+server_variant = "Unsafe"
 
 [clients]
 rust = { Cargo = { cargo_manifest = "oak_functions/examples/mobilenet/client/rust/Cargo.toml" } }
