+++ oak_functions/oak_functions_loader_unsafe/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "oak_functions_loader_unsafe"
+version = "0.1.0"
+edition = "2021"
+license = "Apache-2.0"
+
+[dependencies]
+anyhow = "*"
+oak_functions_loader = { path = "../loader", features = ["oak-unsafe"] }
