+++ oak_functions/oak_functions_loader_unsafe/src/main.rs
@@ -1,0 +1,23 @@
+//
+// Copyright 2022 The Project Oak Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+//! The "unsafe" Oak Functions runtime binary, which does not guarantee that data stays private.
+//!
+//! Useful for testing and debugging, but not meant for production use.
+
+pub fn main() -> anyhow::Result<()> {
+    oak_functions_loader::main()
+}
