+++ scripts/build_reproducibility_index
@@ -12,7 +12,8 @@ source "$SCRIPTS_DIR/common"
 
 # List of artifacts that are expected to be reproducibly built.
 readonly REPRODUCIBLE_ARTIFACTS=(
-  ./oak_functions/loader/bin/oak_functions_loader
+  ./target/x86_64-unknown-linux-musl/release/oak_functions_loader_base
+  ./target/x86_64-unknown-linux-musl/release/oak_functions_loader_unsafe
 )
 
 # Index file containing hashes of the reproducible artifacts, alongside their file names.
