+++ scripts/release_build_oak_functions
@@ -11,4 +11,4 @@ source "$SCRIPTS_DIR/common"
 
 # If the release was built from the latest upstream commit, then this hash should match that from
 # https://github.com/project-oak/oak/blob/hashes/reproducibility_index.
-sha256sum ./oak_functions/loader/bin/oak_functions_loader
+sha256sum ./target/x86_64-unknown-linux-musl/release/oak_functions_loader_base
